Add files via upload
Assignment info:
Now that you have done some work with the tutorial, it is time to apply it to a new page. Create a new folder for this project and call it heropage (all lowercase, all one world). Save the page as index.html.  Use the image below as a guide, and try to replicate it as close as possible using Dreamweaver.  When you are done, create a repository folder called heropage, and upload your html and css to the repository. Share the links for the live site and the repository folder.  The page will be worth 10 total points.  5 points for the accuracy of your page (Are the colors close? Are the fonts similar?  Does the overall layout match? etc.).  The other 5 points are that your files were correctly uploaded to GitHub and you successfully shared the correct links.